query builder: cleanup fetch-structure logic
